requirements.txt
Django==2.2
image==1.5.27
pexif==0.15
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
